requirements.txt
audioread==3.0.0
cairocffi==1.5.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
decorator==4.4.2
docopt==0.6.2
ffmpeg-python==0.2.0
future==0.18.3
gizeh==0.1.11
idna==3.4
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.0.0
joblib==1.2.0
lazy_loader==0.1
librosa==0.10.0.post1
llvmlite==0.39.1
moviepy==1.0.3
msgpack==1.0.5
numba==0.56.4
numpy==1.23.5
packaging==23.0
Pillow==9.4.0
platformdirs==3.1.1
pooch==1.7.0
proglog==0.1.10
pycparser==2.21
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
soundfile==0.12.1
soxr==0.3.4
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarg==0.1.9
zipp==3.15.0
